+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.1 - 2024-03-11
+
+### Fixed
+- Removed guard that would not allow assistant message to be added to request
+- Process response with Jason.decode.
+
+### Improved
+- Added type annotations to documentatio
 ## 0.2.0 - 2024-03-11
 
 ### Improved